Split the 'active span' code from the 'global tracer' code in two packages.
