+++ src/main/java/io/opentracing/contrib/activespan/ThreadLocalSpanManager.java
@@ -1,4 +1,4 @@
-package io.opentracing.contrib.global;
+package io.opentracing.contrib.activespan;
 
 import io.opentracing.Span;
