+++ src/main/java/io/opentracing/contrib/global/concurrent/SpanAwareCallable.java
@@ -1,8 +1,8 @@
 package io.opentracing.contrib.global.concurrent;
 
 import io.opentracing.Span;
-import io.opentracing.contrib.global.ActiveSpanManager;
-import io.opentracing.contrib.global.ActiveSpanManager.SpanDeactivator;
+import io.opentracing.contrib.activespan.ActiveSpanManager;
+import io.opentracing.contrib.activespan.ActiveSpanManager.SpanDeactivator;
 
 import java.util.concurrent.Callable;
