+++ src/main/java/io/opentracing/contrib/global/concurrent/SpanAwareExecutorService.java
@@ -1,7 +1,7 @@
 package io.opentracing.contrib.global.concurrent;
 
 import io.opentracing.Span;
-import io.opentracing.contrib.global.ActiveSpanManager;
+import io.opentracing.contrib.activespan.ActiveSpanManager;
 
 import java.util.ArrayList;
 import java.util.Collection;
