+++ src/main/java/io/opentracing/contrib/global/concurrent/TracedRunnable.java
@@ -1,7 +1,7 @@
 package io.opentracing.contrib.global.concurrent;
 
 import io.opentracing.Span;
-import io.opentracing.contrib.global.ActiveSpanManager;
+import io.opentracing.contrib.activespan.ActiveSpanManager;
 import io.opentracing.contrib.global.GlobalTracer;
 
 import static io.opentracing.contrib.global.concurrent.AddSuppressedSupport.addSuppressedOrLog;
